Extensive fixes and enhancement for `Encode`, `Decode`, and `Transcode`
The functions should now work on MacOS again.
It is now possible to use a custom replacement character in `Encode` and `Transcode`

Added the `GetCpInfo` utility function.

Improved `DecodeUTF8` to bring replacement char insertion more in line with the behaviour of the Windows API function. It is still not 100% identical and more work on that is necessary.