Improvements to `CountSubstring` and `ReplaceFast`
An interesting performance issue with `InStr` used with `vbTextCompare` leads to different algorithms being vastly superior. There findings have been incorporated and taken into account in the affected functions.

A summary of the issue:
In the case of vbTextCompare, InStr's runtime will always be proportional to the find position relative to the beginning of the string, not relative to the specified 'Start' parameter