Another improvement for `LimitConsecutiveSubstringRepetition`
Unfortunately it is impossible to copy the exact behavior of the naive version with this more efficient algorithm as theoretically the order of replacements can always matter for the final outcome. Unfortunately the exact order of replacements of the naive version is extremely difficult to copy in a more efficient manner and that's why in some extremely rare cases the two functions can produce different results. This can only happen when the `subString` partially contains itself, e.g. starts and ends with the same letter, but contains more than one different letter in total and is at least 3 letters long. An example where the order of operations influences the final outcome can look like this:
"aabaaaababa"
Depending on the order of deletion, the result could be either "aab", or "aba"
When trying to limit "aaba" to 0

Because of this, I doubt that the algorithm of this function can be improved much anymore except maybe with a completely different approach.